+ Added image rendering functionality.

+++ workpls/point.cpp
@@ -6,7 +6,7 @@
 		return sqrt(pow(p.x-x,2)+pow(p.y - y, 2));
 	}
 
-	Point::Point(float x, float y)
+	Point::Point(int x, int y)
 	{
 		Point::x = x;
 		Point::y = y;
